+++ backend/src/main/java/org/blog/post/app/backend/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package org.blog.post.app.backend.repository;
+package org.blog.post.app.backend.post.repository;
 
-import org.blog.post.app.backend.model.Post;
+import org.blog.post.app.backend.post.model.Post;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface PostRepository extends MongoRepository<Post, String> {
