+++ backend/src/test/java/org/blog/post/app/backend/service/PostServiceImplTest.java
@@ -1,11 +1,11 @@
 package org.blog.post.app.backend.service;
 
-import org.blog.post.app.backend.dto.PostDTO;
-import org.blog.post.app.backend.exception.ResourceNotFoundException;
-import org.blog.post.app.backend.model.Post;
-import org.blog.post.app.backend.repository.PostRepository;
-import org.blog.post.app.backend.service.impl.PostServiceImpl;
-import org.blog.post.app.backend.util.IdService;
+import org.blog.post.app.backend.post.dto.PostDTO;
+import org.blog.post.app.backend.post.exception.ResourceNotFoundException;
+import org.blog.post.app.backend.post.model.Post;
+import org.blog.post.app.backend.post.repository.PostRepository;
+import org.blog.post.app.backend.post.service.impl.PostServiceImpl;
+import org.blog.post.app.backend.post.util.IdService;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.annotation.DirtiesContext;
